pyproject.toml
[build-system]
requires = ["setuptools>=74", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-ai-reliability-lab"
version = "0.1.0"
description = "A reliability-first clinical-like AI workflow demo with deterministic validation, evaluation loops, and end-to-end observability. Synthetic data only."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Daniel Alvarez" }]
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
  "opentelemetry-api>=1.26",
  "opentelemetry-sdk>=1.26",
  "opentelemetry-exporter-otlp>=1.26",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "opentelemetry-instrumentation-httpx>=0.47b0",
  "reportlab>=4.2",
  "structlog>=24.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "mypy>=1.10",
  "ruff>=0.5",
  "black>=24.4",
  "types-reportlab>=4.2.0.20240611",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "B", "UP", "N", "RUF"]
lint.ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
check_untyped_defs = true
strict_equality = true
mypy_path = ["app"]

[tool.setuptools.packages.find]
include = ["app*"]
